\name{Convert.mgdL.umolL}
\alias{Convert.mgdL.umolL}
\title{
Convertion formula to change from mg/dL to umol/L and vice versa
}
\description{
mg/dL <--> umol/L conversion
}
\usage{
Convert.mgdL.umolL(mg.dL, umol.L)
## If your variable X is expressed in mg/dL, use: Convert.mgdL.mmolL(mg.dL=X)
## If your variable X is expressed in umol/L, use: Convert.mgdL.mmolL(umol.L=X)
}
\arguments{
  \item{mg.dL}{Provide a numeric vector in mg/dL to be converted into umol/L}
  \item{umol.L}{Provide a numeric vector in umol/L to be converted into mg/dL}
}
\value{
A numeric vector with the converted values.
}
\author{
Cristian Pattaro
}

\seealso{
\code{\link[nephro]{Convert.mgdL.umolL}}
}
\keyword{ conversion }
\keyword{ creatinine }
